Use positional arguments for repeat() and loosen signature
Put 'outer' first as it is more natural to omit 'inner'.
Deprecate the old version with keyword arguments.
Also change the signature to accept any AbstractArray as
first argument, and any type for 'outer' and 'inner', to
allow in particular passing tuples or arrays. Allow
passing an empty collection to imply a no-op.